+++ packages/live-preview-sdk/src/helpers/validation.ts
@@ -1,6 +1,8 @@
 import { DocumentNode } from 'graphql';
 import gql from 'graphql-tag';
+
 import type { ContentfulSubscribeConfig } from '..';
+
 import { Argument, SubscribeCallback } from '../types';
 import { debug } from './debug';