+++ packages/live-preview-sdk/src/inspectorMode/utils.ts
@@ -1,4 +1,30 @@
+import { decode } from '../csm/encode';
 import { InspectorModeAttributes, InspectorModeDataAttributes } from './types';
+
+const isTaggedElement = (node?: Node | null): boolean => {
+  if (!node) {
+    return false;
+  }
+
+  if (node.nodeType !== Node.ELEMENT_NODE) {
+    return false;
+  }
+
+  const el = node as Element;
+
+  if (!el.hasAttribute(InspectorModeDataAttributes.FIELD_ID)) {
+    return false;
+  }
+
+  if (
+    !el.hasAttribute(InspectorModeDataAttributes.ENTRY_ID) &&
+    !el.hasAttribute(InspectorModeDataAttributes.ASSET_ID)
+  ) {
+    return false;
+  }
+
+  return true;
+};
 
 /**
  * Parses the necessary information from the element and returns them.
@@ -6,19 +32,24 @@ import { InspectorModeAttributes, InspectorModeDataAttributes } from './types';
  */
 export function getInspectorModeAttributes(
   element: Element,
-  fallbackLocale?: string
+  fallbackLocale: string
 ): InspectorModeAttributes | null {
-  const fieldId = element.getAttribute(InspectorModeDataAttributes.FIELD_ID);
-  const entryId = element.getAttribute(InspectorModeDataAttributes.ENTRY_ID);
-  const assetId = element.getAttribute(InspectorModeDataAttributes.ASSET_ID);
-  const locale = element.getAttribute(InspectorModeDataAttributes.LOCALE) ?? fallbackLocale;
-
-  if (assetId && locale && fieldId) {
-    return { fieldId, assetId, locale };
+  if (!isTaggedElement(element)) {
+    return null;
   }
 
-  if (entryId && locale && fieldId) {
-    return { fieldId, entryId, locale };
+  const fieldId = element.getAttribute(InspectorModeDataAttributes.FIELD_ID) as string;
+  const locale = element.getAttribute(InspectorModeDataAttributes.LOCALE) ?? fallbackLocale;
+
+  const entryId = element.getAttribute(InspectorModeDataAttributes.ENTRY_ID);
+  const assetId = element.getAttribute(InspectorModeDataAttributes.ASSET_ID);
+
+  if (entryId) {
+    return { entryId, fieldId, locale };
+  }
+
+  if (assetId) {
+    return { assetId, fieldId, locale };
   }
 
   return null;
@@ -26,15 +57,79 @@ export function getInspectorModeAttributes(
 
 /**
  * Query the document for all tagged elements
- * **Attention:** Can include elements that have not all attributes,
- * if you want to have only valid ones check for `getTaggedInformation`
  */
-export function getAllTaggedElements(): Element[] {
-  const allElements = [
-    ...document.querySelectorAll(`[${InspectorModeDataAttributes.ENTRY_ID}]`),
-    ...document.querySelectorAll(`[${InspectorModeDataAttributes.ASSET_ID}]`),
-  ];
-  return allElements;
+export function getAllTaggedElements(root = window.document, ignoreManual?: boolean): Element[] {
+  // The fastest way to look up & iterate over DOM. Ref:
+  // https://stackoverflow.com/a/2579869
+  //
+  // Terminology:
+  // FILTER_SKIP: Skip the current node
+  // FILTER_REJECT: Skip the current node and all its children
+  const walker = document.createTreeWalker(root, NodeFilter.SHOW_ALL, (node) => {
+    if (node.nodeType === Node.TEXT_NODE) {
+      const text = node.textContent ?? '';
+
+      const decoded = decode(text);
+
+      if (decoded?.origin !== 'contentful.com') {
+        return NodeFilter.FILTER_SKIP;
+      }
+
+      return isTaggedElement(node.parentElement)
+        ? ignoreManual
+          ? NodeFilter.FILTER_ACCEPT
+          : NodeFilter.FILTER_SKIP
+        : NodeFilter.FILTER_ACCEPT;
+    }
+
+    if (ignoreManual) {
+      return NodeFilter.FILTER_SKIP;
+    }
+
+    return isTaggedElement(node) ? NodeFilter.FILTER_ACCEPT : NodeFilter.FILTER_SKIP;
+  });
+
+  const elements: Element[] = [];
+
+  while (walker.nextNode()) {
+    const node = walker.currentNode;
+
+    if (node.nodeType === Node.ELEMENT_NODE) {
+      elements.push(walker.currentNode as Element);
+      continue;
+    }
+
+    if (!node.textContent) {
+      continue;
+    }
+    // Handle Encoded strings
+    const decoded = decode(node.textContent);
+
+    if (!decoded?.contentful) {
+      continue;
+    }
+
+    const { contentful } = decoded;
+    const el = node.parentElement;
+
+    if (!el) {
+      continue;
+    }
+
+    if (contentful.entityType === 'Entry') {
+      el.setAttribute(InspectorModeDataAttributes.ENTRY_ID, contentful.entity);
+    } else {
+      el.setAttribute(InspectorModeDataAttributes.ASSET_ID, contentful.entity);
+    }
+
+    // TODO: add space/env ids to properly handle cross-space content
+    el.setAttribute(InspectorModeDataAttributes.LOCALE, contentful.locale);
+    el.setAttribute(InspectorModeDataAttributes.FIELD_ID, contentful.field);
+
+    elements.push(el);
+  }
+
+  return elements;
 }
 
 /**